refactor(auth): Se adaptó el servicio de auth a los cambios en la db.
