+++ backend/prisma/schema.prisma
@@ -187,7 +187,7 @@ model WebSession {
   ip             String   @db.VarChar(45)
   userAgent      String   @map("user_agent") @db.VarChar(512)
   active         Boolean  @map("activo") @db.TinyInt
-  createdAt      DateTime @map("fecha_creacion")
+  createdDate    DateTime @map("fecha_creacion")
   expirationDate DateTime @map("fecha_expiracion")
 
   webUser WebUser @relation(fields: [userId], references: [id], onDelete: Restrict, onUpdate: Restrict, map: "sesiones_web_ibfk_1")
@@ -196,12 +196,12 @@ model WebSession {
 }
 
 model WebUser {
-  id         String @id @db.VarChar(36)
-  dni        String @unique @db.VarChar(20)
-  avatarHash String @map("avatar_hash") @db.VarChar(128)
-  email      String @map("correo_electronico") @db.VarChar(255)
-  password   String @map("contrasena") @db.VarChar(255)
-  admin      Int    @default(0) @db.TinyInt
+  id         String  @id @db.VarChar(36)
+  dni        String  @unique @db.VarChar(20)
+  avatarHash String  @map("avatar_hash") @db.VarChar(128)
+  email      String  @unique @map("correo_electronico") @db.VarChar(255)
+  password   String  @map("contrasena") @db.VarChar(255)
+  admin      Boolean @default(false) @db.TinyInt
 
   user     User         @relation(fields: [dni], references: [dni], onDelete: Restrict, onUpdate: Restrict, map: "FK_usuarios_web_usuarios")
   sessions WebSession[]
